+++ Runtime/TexTransMonoBase.cs
@@ -36,6 +36,7 @@ namespace net.rs64.TexTransTool
         //v0.8.x == 5
         //v0.9.x == 6
         //v0.10.x == 7
+        //v1.0.x == 7
         internal const int TTTDataVersion_0_10_X = 7;
         internal const int TTTDataVersion = 7;
 
@@ -44,6 +45,21 @@ namespace net.rs64.TexTransTool
 
 
         internal void OnDestroy() => MonoBehaviourCallProvider.DestroyThis(this);
+
+        internal static bool IsOldSaveData(TexTransMonoBase monoBase)
+        {
+            if (monoBase is ITexTransToolStableComponent texTransToolStableComponent)
+            {
+                if ((monoBase as ITexTransToolTag).SaveDataVersion < texTransToolStableComponent.StabilizeSaveDataVersion)
+                    return true;
+            }
+            else
+            {
+                if ((monoBase as ITexTransToolTag).SaveDataVersion < TexTransMonoBase.TTTDataVersion)
+                    return true;
+            }
+            return false;
+        }
     }
     [DisallowMultipleComponent]
     public abstract class TexTransMonoBaseGameObjectOwned : TexTransMonoBase { }